refactor(iommu_map_pa): make the wrapper behave like the importer
The userspace wrapper lived inside the DKMS package source, so a consumer of
an installed uPCIe could not include it, and it resolved a missing UAPI by
reaching sideways into the module's own copy through a relative path. That
path only exists in a full checkout: vendored into another project, as xNVMe
vendors include/upcie/, the include fails outright.

So put it where the importer wrapper goes, install it from there, and give it
the same shape: guard on the UAPI, expose UPCIE_HAVE_IOMMU_MAP_PA, and compile
to -ENOTSUP stubs when the package is absent. Two experimental modules that
behave the same way are one thing to learn rather than two. That also makes it
safe in upcie.h, and lets the CI consumer assert it built the real path rather
than passing on stubs.

Matching it properly means two more things. The helpers drop their upcie_
prefix, since the point of the rename is that nothing here is upcie-specific
and the sibling spells its own dmabuf_import_attach() without one. And the
stub branch defines the constants its arguments are built from: they live only
in the UAPI, so a caller passing IOMMU_MAP_PA_PROT_READ could not compile
without the package, which is the situation the stubs exist to avoid.

The signatures lose __u32 and __u64 for their stdint spellings, since the stub
path has no kernel headers to define them.
